ARGO_CHATBOT/requirements.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
tkintermapview
matplotlib
pandas
requests
python-dotenv
pyautogui
pillow
